Restructure Vec classes into one file
